feat: Adjust component WhoWeAre folder structure
- Move avatar to `@/public/images`
- Extract roleDisplayTextMap logic to `@/constants`
- Put WhoWeAre section constant data to the component
