--- nextjs-app/components/pages/main/WhoWeAre/index.ts
@@ -1,1 +1,0 @@
-export { default } from "./WhoWeAre";
