+++ nextjs-app/constants/roleDisplayTextMap.ts
@@ -1,0 +1,8 @@
+import { Role } from "@/types";
+
+export const roleDisplayTextMap: Record<Role, string> = {
+  [Role.CAPTAIN]: "船長",
+  [Role.HARBOUR_PILOT]: "引水人",
+  [Role.NAVIGATOR]: "航海士",
+  [Role.SAILOR]: "水手",
+};
